remove UpdateState function Prior to this change, UpdateState was really only used when initializing sauced. This was going to be renamed to InitState but this is not needed.
This is because InitState did two things:

1. Went over the LastKnownState and update that based on the parsed metadata file
2. Created the sauced-state.json file if it wasn't created.

For 1, this is something we want to implement but at the moment it is not doing what it is supposed to.
Since this is only called on initialization, there is nothing to update and it never even enters the loop
where things get marshalled/updated.

For 2, because of the behavior of WriteFile it is not needed.
If for any reason, the file does not exist, the file will be created automatically.